move more data stuff in data
